krypto-backend/requirements.txt
Flask
flask-cors
Pillow
pytesseract
